feat: share EvalHub tenant fixtures and add MCP auth success test
Reuse parent evalhub conftest fixtures for tenant RBAC and vLLM emulator
across MCP and multitenancy tests, and add an explicit positive auth test
that verifies token creation and proxy RBAC allow MCP initialize.
