+++ tests/ai_safety/evalhub/conftest.py
@@ -31,16 +31,19 @@ from tests.ai_safety.evalhub.constants import (
     EVALHUB_JOB_SA_SUFFIX,
     EVALHUB_JOBS_WRITER_CLUSTERROLE,
     EVALHUB_METRICS_SERVICE_SUFFIX,
+    EVALHUB_MT_CR_NAME,
     EVALHUB_TENANT_LABEL_KEY,
     EVALHUB_TENANT_LABEL_VALUE,
     EVALHUB_USER_ROLE_RULES,
+    EVALHUB_VLLM_EMULATOR_PORT,
     GARAK_INTENTS_S3_KEY,
     MINIO_MC_IMAGE,
     MINIO_UPLOADER_SECURITY_CONTEXT,
 )
-from tests.ai_safety.evalhub.utils import wait_for_service_account
+from tests.ai_safety.evalhub.kueue.constants import VLLM_EMULATOR, VLLM_EMULATOR_IMAGE
+from tests.ai_safety.evalhub.utils import tenant_rbac_ready, wait_for_service_account
 from utilities.certificates_utils import create_ca_bundle_file
-from utilities.constants import Timeout
+from utilities.constants import Labels, Protocols, Timeout
 from utilities.general import collect_pod_information
 from utilities.infra import create_inference_token, create_ns
 
@@ -160,6 +163,118 @@ def evalhub_mt_ca_bundle_file(
 ) -> str:
     """CA bundle file for verifying TLS on the EvalHub route."""
     return create_ca_bundle_file(client=admin_client)
+
+
+@pytest.fixture(scope="class")
+def evalhub_tenant_rbac_instance_name() -> str:
+    """EvalHub CR name used when waiting for operator job RBAC in tenant namespaces."""
+    return EVALHUB_MT_CR_NAME
+
+
+@pytest.fixture(scope="class")
+def evalhub_tenant_deployment(evalhub_mt_deployment: Deployment) -> Deployment:
+    """EvalHub deployment whose operator RBAC must be ready in tenant namespaces."""
+    return evalhub_mt_deployment
+
+
+@pytest.fixture(scope="class")
+def tenant_a_rbac_ready(
+    admin_client: DynamicClient,
+    tenant_a_namespace: Namespace,
+    evalhub_tenant_deployment: Deployment,
+    evalhub_tenant_rbac_instance_name: str,
+) -> None:
+    """Wait for the operator to provision job RBAC in tenant-a."""
+    try:
+        for ready in TimeoutSampler(
+            wait_timeout=120,
+            sleep=5,
+            func=tenant_rbac_ready,
+            admin_client=admin_client,
+            namespace=tenant_a_namespace.name,
+            evalhub_instance_name=evalhub_tenant_rbac_instance_name,
+        ):
+            if ready:
+                LOGGER.info(f"Operator RBAC provisioned in {tenant_a_namespace.name}")
+                return
+    except TimeoutExpiredError as err:
+        msg = (
+            f"Operator RBAC provision failed: RoleBindings, ServiceAccount, or service-CA ConfigMap"
+            f" not found in namespace '{tenant_a_namespace.name}' within timeout"
+        )
+        LOGGER.error(msg)
+        raise RuntimeError(msg) from err
+
+
+@pytest.fixture(scope="class")
+def evalhub_vllm_emulator_deployment(
+    admin_client: DynamicClient,
+    tenant_a_namespace: Namespace,
+    tenant_a_rbac_ready: None,
+) -> Generator[Deployment, Any, Any]:
+    """Deploy the vLLM emulator in tenant-a for job submission tests."""
+    label = {Labels.Openshift.APP: VLLM_EMULATOR}
+    with Deployment(
+        client=admin_client,
+        namespace=tenant_a_namespace.name,
+        name=VLLM_EMULATOR,
+        label=label,
+        selector={"matchLabels": label},
+        template={
+            "metadata": {
+                "labels": label,
+                "name": VLLM_EMULATOR,
+            },
+            "spec": {
+                "containers": [
+                    {
+                        "image": VLLM_EMULATOR_IMAGE,
+                        "name": VLLM_EMULATOR,
+                        "ports": [{"containerPort": EVALHUB_VLLM_EMULATOR_PORT, "protocol": Protocols.TCP}],
+                        "readinessProbe": {
+                            "tcpSocket": {"port": EVALHUB_VLLM_EMULATOR_PORT},
+                            "initialDelaySeconds": 5,
+                            "periodSeconds": 5,
+                            "timeoutSeconds": 3,
+                            "failureThreshold": 6,
+                        },
+                        "securityContext": {
+                            "allowPrivilegeEscalation": False,
+                            "capabilities": {"drop": ["ALL"]},
+                            "seccompProfile": {"type": "RuntimeDefault"},
+                        },
+                    }
+                ]
+            },
+        },
+        replicas=1,
+    ) as deployment:
+        deployment.wait_for_replicas(timeout=Timeout.TIMEOUT_5MIN)
+        yield deployment
+
+
+@pytest.fixture(scope="class")
+def evalhub_vllm_emulator_service(
+    admin_client: DynamicClient,
+    tenant_a_namespace: Namespace,
+    evalhub_vllm_emulator_deployment: Deployment,
+) -> Generator[Service, Any, Any]:
+    """Service fronting the vLLM emulator in tenant-a."""
+    with Service(
+        client=admin_client,
+        namespace=tenant_a_namespace.name,
+        name=f"{VLLM_EMULATOR}-service",
+        ports=[
+            {
+                "name": f"{VLLM_EMULATOR}-endpoint",
+                "port": EVALHUB_VLLM_EMULATOR_PORT,
+                "protocol": Protocols.TCP,
+                "targetPort": EVALHUB_VLLM_EMULATOR_PORT,
+            }
+        ],
+        selector={Labels.Openshift.APP: VLLM_EMULATOR},
+    ) as service:
+        yield service
 
 
 @pytest.fixture(scope="class")
