+++ tests/ai_safety/evalhub/mcp/test_evalhub_mcp_auth.py
@@ -2,7 +2,7 @@ import pytest
 from ocp_resources.namespace import Namespace
 from ocp_resources.route import Route
 
-from tests.ai_safety.evalhub.mcp.utils import EvalHubMcpClient
+from tests.ai_safety.evalhub.mcp.utils import EvalHubMcpClient, validate_evalhub_mcp_initialize
 from tests.ai_safety.evalhub.utils import TENANT_HEADER, build_headers
 from utilities.guardrails import get_auth_headers
 
@@ -21,6 +21,26 @@ from utilities.guardrails import get_auth_headers
 @pytest.mark.usefixtures("evalhub_mcp_mt_ready", "evalhub_mcp_proxy_role_binding")
 class TestEvalHubMcpAuth:
     """Authentication tests for evalhub-mcp behind kube-rbac-proxy."""
+
+    def test_mcp_request_with_valid_token_succeeds(
+        self,
+        tenant_a_token: str,
+        tenant_a_namespace: Namespace,
+        evalhub_mcp_mt_route: Route,
+        evalhub_mcp_mt_ca_bundle_file: str,
+    ) -> None:
+        """
+        Given: A valid bearer token, X-Tenant header, and evalhubs/proxy RBAC
+        When: MCP initialize request is sent
+        Then: Server accepts the request and returns initialize result
+        """
+        result = validate_evalhub_mcp_initialize(
+            host=evalhub_mcp_mt_route.host,
+            token=tenant_a_token,
+            ca_bundle_file=evalhub_mcp_mt_ca_bundle_file,
+            tenant_namespace=tenant_a_namespace.name,
+        )
+        assert result.get("serverInfo"), "Expected serverInfo in initialize result"
 
     def test_mcp_request_without_token_is_rejected(
         self,
