+++ tests/ai_safety/evalhub/mcp/test_evalhub_mcp_jobs.py
@@ -34,7 +34,7 @@ class TestEvalHubMcpJobs:
     def test_get_job_status_returns_progress_fields(
         self,
         evalhub_mcp_client: EvalHubMcpClient,
-        evalhub_mcp_vllm_emulator_service: Service,
+        evalhub_vllm_emulator_service: Service,
         tenant_a_namespace: Namespace,
     ) -> None:
         """
@@ -43,7 +43,7 @@ class TestEvalHubMcpJobs:
         Then: Response includes job_id, state, and progress_percent
         """
         model_url = build_mcp_model_url(
-            service_name=evalhub_mcp_vllm_emulator_service.name,
+            service_name=evalhub_vllm_emulator_service.name,
             tenant_namespace=tenant_a_namespace.name,
         )
         submit_result = submit_evaluation_via_mcp(
@@ -69,7 +69,7 @@ class TestEvalHubMcpJobs:
     def test_cancel_running_job(
         self,
         evalhub_mcp_client: EvalHubMcpClient,
-        evalhub_mcp_vllm_emulator_service: Service,
+        evalhub_vllm_emulator_service: Service,
         tenant_a_namespace: Namespace,
     ) -> None:
         """
@@ -78,7 +78,7 @@ class TestEvalHubMcpJobs:
         Then: Job reaches cancelled terminal state
         """
         model_url = build_mcp_model_url(
-            service_name=evalhub_mcp_vllm_emulator_service.name,
+            service_name=evalhub_vllm_emulator_service.name,
             tenant_namespace=tenant_a_namespace.name,
         )
         submit_result = submit_evaluation_via_mcp(
